requirements.txt
altair
pandas
streamlit
streamlit_chat
streamlit_javascript
numpy
pillow
sentence-transformers